Move word controller logic into the routes
Done to be able to harness type inference more easily. Resulted in quite
a bloated file, but there is a lot of repetitive code that can be
abstracted into a helper function or fixed in some other way.
